--- a/.husky/pre-commit
+++ b/.husky/pre-commit
@@ -1,4 +1,4 @@
 pnpm run lint
 pnpm generateAutomd
-prettier $(git diff --cached --name-only --diff-filter=ACMR | sed 's| |\\ |g') --write --ignore-unknown
+biome format --write --no-errors-on-unmatched $(git diff --cached --name-only --diff-filter=ACMR | sed 's| |\\ |g')
 git update-index --again