--- a/packages/cms-base/components.d.ts
+++ b/packages/cms-base/components.d.ts
@@ -1,7 +1,7 @@
 declare module "@vue/runtime-core" {
   export type GlobalComponents = {
-    CmsGenericBlock: (typeof import("./components/cms/CmsGenericBlock.vue"))["default"];
-    CmsGenericElement: (typeof import("./components/cms/CmsGenericElement.vue"))["default"];
+    CmsGenericBlock: typeof import("./components/cms/CmsGenericBlock.vue")["default"];
+    CmsGenericElement: typeof import("./components/cms/CmsGenericElement.vue")["default"];
   };
 
   declare interface AllowedComponentProps {