--- a/packages/cms-base/components.d.ts
+++ b/packages/cms-base/components.d.ts
@@ -9,4 +9,4 @@ declare module "@vue/runtime-core" {
   }
 }
 
-export {};
+export type {};