--- a/templates/astro/src/layouts/Layout.astro
+++ b/templates/astro/src/layouts/Layout.astro
@@ -1,6 +1,6 @@
 ---
 export interface Props {
-	title: string;
+  title: string;
 }
 
 const { title } = Astro.props;