--- a/templates/astro/src/pages/index.astro
+++ b/templates/astro/src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
-import Layout from '../layouts/Layout.astro';
-import Session from '../components/Session.vue';
+import Layout from "../layouts/Layout.astro";
+import Session from "../components/Session.vue";
 ---
 
 <Layout title="Shopware Frontends & Astro">