--- a/templates/vue-demo-store/app.vue
+++ b/templates/vue-demo-store/app.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import { getPrefix } from "./i18n/src/helpers/prefix";
 import type { Schemas } from "#shopware";
+import { getPrefix } from "./i18n/src/helpers/prefix";
 const config = useRuntimeConfig();
 /**
  * Init breadcrumbs context