--- a/templates/vue-demo-store/app/router.options.ts
+++ b/templates/vue-demo-store/app/router.options.ts
@@ -32,11 +32,10 @@ export default <RouterOptions>{
 
     if (savedPosition) {
       return savedPosition;
-    } else {
-      return {
-        top: 0,
-        behavior: "smooth",
-      };
     }
+    return {
+      top: 0,
+      behavior: "smooth",
+    };
   },
 };