--- a/templates/vue-demo-store/components/account/AccountPersonalData.vue
+++ b/templates/vue-demo-store/components/account/AccountPersonalData.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import { useVuelidate } from "@vuelidate/core";
 import { customValidators } from "@/i18n/utils/i18n-validators";
 import { ApiClientError, type ApiError } from "@shopware/api-client";
+import { useVuelidate } from "@vuelidate/core";
 const { required, minLength, requiredIf, email, sameAs } = customValidators();
 
 const { user, refreshUser, updatePersonalInfo, updateEmail } = useUser();