--- a/templates/vue-demo-store/components/checkout/CheckoutCartItem.vue
+++ b/templates/vue-demo-store/components/checkout/CheckoutCartItem.vue
@@ -14,11 +14,10 @@ const props = withDefaults(
 const { cartItem } = toRefs(props);
 const CartItem = computed(() => {
   switch (cartItem.value?.type) {
-    default:
-    case "product":
-      return defineAsyncComponent(() => import("./cart/ProductItem.vue"));
     case "promotion":
       return defineAsyncComponent(() => import("./cart/PromotionItem.vue"));
+    default:
+      return defineAsyncComponent(() => import("./cart/ProductItem.vue"));
   }
 });
 </script>