--- a/templates/vue-demo-store/components/layout/LayoutSideMenu.vue
+++ b/templates/vue-demo-store/components/layout/LayoutSideMenu.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import {
-  getCategoryRoute,
   getCategoryImageUrl,
+  getCategoryRoute,
 } from "@shopware-pwa/helpers-next";
 import type { Schemas } from "#shopware";