--- a/templates/vue-demo-store/components/listing-filters/ListingFiltersShippingFree.vue
+++ b/templates/vue-demo-store/components/listing-filters/ListingFiltersShippingFree.vue
@@ -10,17 +10,18 @@
     }
   "
 >
-import type { Schemas } from "#shopware";
 import { computed } from "vue";
+import type { Schemas } from "#shopware";
 
 const props = defineProps<{
   filter: ListingFilter;
   selectedFilters: Schemas["ProductListingResult"]["currentFilters"];
 }>();
 
-const emits = defineEmits<{
-  (e: "select-value", value: { code: string; value: unknown }): void;
-}>();
+const emits =
+  defineEmits<
+    (e: "select-value", value: { code: string; value: unknown }) => void
+  >();
 const currentFilterData = computed(
   () => !!props.selectedFilters[props.filter?.code],
 );
@@ -37,7 +38,9 @@ const toggle = () => {
 };
 
 const dropdownElement = ref(null);
-onClickOutside(dropdownElement, () => (isFilterVisible.value = false));
+onClickOutside(dropdownElement, () => {
+  isFilterVisible.value = false;
+});
 </script>
 
 <template>