--- a/templates/vue-demo-store/components/shared/SharedPagination.vue
+++ b/templates/vue-demo-store/components/shared/SharedPagination.vue
@@ -4,9 +4,7 @@ defineProps<{
   current: number;
 }>();
 
-defineEmits<{
-  (e: "changePage", page: number): void;
-}>();
+defineEmits<(e: "changePage", page: number) => void>();
 </script>
 <template>
   <nav