@patzick 2024-12-03
sidenote: needed to move biome back to 1.8.3 and turn off IDE plugin until https://github.com/biomejs/biome/issues/4684 is resolved. As that causes problem in `shopware.d.ts` files